bot/requirements.txt
BeatPy
discord.py
